Remove TERM=xterm-256color in bashrc
That setting in no longer needed since it must be configured by the
terminal emulator application. In the case of MobaXterm, when creating a
new session or editing an existing one, go to "Terminal settings" and
select the appropriate "Terminal type". In my case, the one that I
selected is xterm-256color, just please be sure to verify that the
terminfo entry for that particular terminal type is already installed in
your system with the infocmp command:

infocmp xterm-256color